Unlink InCallUI background from Sim colors
This commit unlinks the InCallUI background from the sim colors
located under Settings/Network & Internet/SIM cards. It now changes
based on the system accent as do other things within our ROM.

We're also keeping themers in mind with this commit. We know that many
themers like to do their own thing here. If a themer wants to default
back to the AOSP behavior, all that's needs is a simple change of a bool.

'config_has_theme_applied' needs to be set to true for themers

Originally this setting the accent programmatically and checking for the accent
overlay but that proved to be more than this package can handle. So after further
review, I decided to keep this simple.

Screenshots:
- https://i.imgur.com/cD0eY5q.jpg
- https://i.imgur.com/ZlRwFwW.jpg
- https://i.imgur.com/p3qOnOT.png
- https://i.imgur.com/0z5ZHo3.jpg

Change-Id: I3e15ca48b27be0e70709d2ef60ffbb1b4a8c7b3a
Signed-off-by: spezi77 <spezi7713@gmx.net>